alerts: Add concept of "stabilisation"
Alerts now have three states: active, stabilising, and inactive. Alerts
become active when the threshold is breached, stabilising when the alert
goes under the threshold again, and inactive after expiry_sec seconds
are elapsed (at the granularity of one interval).

This is one of the ways we can reduce volatility, as discussed in #16.