alerts: Add 5% grace period to reduce volatility
This is part of the volatility reduction work in #16. At 5% below a
threshold up to the current threshold, we will not touch the alert's
current state at all. This reduces the thrash when there is a workload
which is bouncing pressures up and down within a small range close to
the defined threshold repeatedly.